Handle all chat actions with the new floating widget.
