+++ src/vs/workbench/contrib/notebook/browser/controller/chat/notebook.chat.contribution.ts
@@ -1,0 +1,6 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import 'vs/workbench/contrib/notebook/browser/controller/chat/cellChatActions';
