+++ src/vs/workbench/contrib/notebook/browser/notebook.contribution.ts
@@ -67,6 +67,7 @@ import 'vs/workbench/contrib/notebook/browser/controller/editActions';
 import 'vs/workbench/contrib/notebook/browser/controller/cellOutputActions';
 import 'vs/workbench/contrib/notebook/browser/controller/apiActions';
 import 'vs/workbench/contrib/notebook/browser/controller/foldingController';
+import 'vs/workbench/contrib/notebook/browser/controller/chat/notebook.chat.contribution';
 
 // Editor Contribution
 import 'vs/workbench/contrib/notebook/browser/contrib/editorHint/emptyCellEditorHint';
